Feature/modify account (#94)
* Created Unlinked ModifyAccount page logic

Still needs to be tested, so I'll add the buttons on the dashboard to connect to this page and test it later.

* Hooked up dashboard to modify account page

* Completed Modify Account Page

Also cleaned up the UI over-all and fixed some of the graphical display issues (pointless menu buttons etc)

* Made UI more consistent

Colors and text should now be consistent across pages. Code consistency will be fixed in the next commit.