Dev/android (#100)
* initialized the android project

* login android pages update

* signup update

* login and signup page v1

* created modify_account.xml and new constants

* Updated button colors and formatting

* Created Dashboard UI (#92)

Mapped the logout buttons. Uncomment and complete the routing for the other buttons once those pages are created.

* Feature/modify account (#93)

* Created Unlinked ModifyAccount page logic

Still needs to be tested, so I'll add the buttons on the dashboard to connect to this page and test it later.

* Hooked up dashboard to modify account page

* Completed Modify Account Page

Also cleaned up the UI over-all and fixed some of the graphical display issues (pointless menu buttons etc)

* Feature/modify account (#94)

* Created Unlinked ModifyAccount page logic

Still needs to be tested, so I'll add the buttons on the dashboard to connect to this page and test it later.

* Hooked up dashboard to modify account page

* Completed Modify Account Page

Also cleaned up the UI over-all and fixed some of the graphical display issues (pointless menu buttons etc)

* Made UI more consistent

Colors and text should now be consistent across pages. Code consistency will be fixed in the next commit.

* Reformatted code and added proper error checking

* Added Contact Info (#95)

Added a new contact info page accessible from the main page.
Also removed Assistant Login button as it was never supposed to be supported.

* Feature/view tech sched (#96)

* Started work on viewing timeslots

* Mostly finished timeslot view

* Displays workspace name for tiemslot view

* Added comments...

* Renamed duplicate string resource

* booking layout

* adjustments in layout

* book button

* timeslot container implemented

* Fixed book appointment button

* Addition of the ViewAppointments Page (#97)

* early draft

* draft2

* draft3 (restart) displays type and id as test

* added switch

* Displaying appointments (without decorations)

* Tentative Final Version of ViewAppointments Page

* clean up

* Update strings.xml

* Updated title text in book appointment page

Co-authored-by: Invincidude <adel.ahram2@gmail.com>

* Auto Code Formatting and final cleanup

Co-authored-by: Alix <alix.routhier-lalonde@hotmail.com>
Co-authored-by: Roey Borsteinas <roeyb1@hotmail.com>
Co-authored-by: Simon <simon.nakanemarcil@mail.mcgill.ca>
Co-authored-by: curious-duck <59633108+curious-duck@users.noreply.github.com>